+++ src/main/java/com/tale/utils/TaleUtils.java
@@ -13,6 +13,7 @@ import com.sun.syndication.feed.rss.Item;
 import com.sun.syndication.io.FeedException;
 import com.sun.syndication.io.WireFeedOutput;
 import com.tale.ext.Commons;
+import com.tale.ext.Theme;
 import com.tale.init.TaleConst;
 import com.tale.model.Contents;
 import com.tale.model.Users;
@@ -242,9 +243,9 @@ public class TaleUtils {
             Item item = new Item();
             item.setTitle(post.getTitle());
             Content content = new Content();
-            content.setValue(Commons.article(post.getContent()));
+            content.setValue(Theme.article(post.getContent()));
             item.setContent(content);
-            item.setLink(Commons.permalink(post.getCid(), post.getSlug()));
+            item.setLink(Theme.permalink(post.getCid(), post.getSlug()));
             item.setPubDate(DateKit.getDateByUnixTime(post.getCreated()));
             items.add(item);
         });
