+++ src/main/resources/templates/themes/default/archives.html
@@ -10,10 +10,10 @@
                     <div class="post-list-item-container">
                         <div class="item-label">
                             <div class="item-title">
-                                <a href="${(permalink(article.cid, article.slug))}">${article.title}</a>
+                                <a href="${(permalink())}">${title()}</a>
                             </div>
                             <div class="item-meta clearfix">
-                                <div class="item-meta-date">发布于 ${fmtdate(article.created)}</div>
+                                <div class="item-meta-date">发布于 ${created('yyyy-MM-dd')}</div>
                             </div>
                         </div>
                     </div>
