+++ src/main/resources/templates/themes/default/index.html
@@ -5,20 +5,20 @@
             #for(article : articles.list)
             <div class="post-list-item">
                 <div class="post-list-item-container">
-                    <a href="${permalink(article)}" class="item-thumb  bg-deepgrey"
+                    <a href="${permalink()}" class="item-thumb  bg-deepgrey"
                        style="background-image:url(${show_thumb(article)});">
-                        <dir class="item-desc">${intro(article.content, 75)}...</dir>
+                        <dir class="item-desc">${intro(75)}...</dir>
                     </a>
                     <div class="item-slant reverse-slant &lt; bg-deepgrey"></div>
                     <div class="item-slant"></div>
                     <div class="item-label">
-                        <div class="item-title"><a href="${permalink(article)}">${article.title}</a>
+                        <div class="item-title"><a href="${permalink()}">${title()}</a>
                         </div>
                         <div class="item-meta clearfix">
-                            <div class="item-meta-ico ${show_icon(article.cid)}"
+                            <div class="item-meta-ico ${show_icon()}"
                                  style="background: url(/static/user/img/bg-ico.png) no-repeat;background-size: 40px auto;"></div>
                             <div class="item-meta-cat">
-                                ${show_categories(article.categories)}
+                                ${show_categories()}
                             </div>
                         </div>
                     </div>
