+++ src/main/resources/templates/themes/default/page.html
@@ -1,12 +1,12 @@
 #include('./header.html', {title: article.title })
 <article class="main-content post-page">
     <div class="post-header">
-        <h1 class="post-title" itemprop="name headline">${article.title}</h1>
+        <h1 class="post-title" itemprop="name headline">${title()}</h1>
         <div class="post-data">
-            <time datetime="${fmtdate(article.created)}" itemprop="datePublished">发布于 ${fmtdate(article.created)}</time>
+            <time datetime="${created('yyyy-MM-dd')}" itemprop="datePublished">发布于 ${created('yyyy-MM-dd')}</time>
         </div>
     </div>
-    <div id="post-content" class="post-content">${article(article.content)}</div>
+    <div id="post-content" class="post-content">${show_content()}</div>
 </article>
 #include('./comments.html')
 #include('./footer.html')
